Move `setupDefaultChips` to a seperate file.
