+++ assets/js/components/DashboardNavigation/index.test.js
@@ -26,69 +26,9 @@ import {
 } from '../../googlesitekit/constants';
 import { CORE_MODULES } from '../../googlesitekit/modules/datastore/constants';
 import { CORE_WIDGETS } from '../../googlesitekit/widgets/datastore/constants';
-import {
-	CONTEXT_MAIN_DASHBOARD_CONTENT,
-	CONTEXT_MAIN_DASHBOARD_SPEED,
-	CONTEXT_MAIN_DASHBOARD_TRAFFIC,
-} from '../../googlesitekit/widgets/default-contexts';
+import { CONTEXT_MAIN_DASHBOARD_SPEED } from '../../googlesitekit/widgets/default-contexts';
 import DashboardNavigation from './';
-
-/**
- * Dispatches required actions to registry to make sure widget contexts for Traffic, Content & Speed are active.
- *
- * @since 1.47.0
- *
- * @param {Object} registry The registry object.
- */
-export const setupDefaultChips = ( registry ) => {
-	// Traffic
-	registry.dispatch( CORE_WIDGETS ).registerWidgetArea( 'TrafficArea', {
-		title: 'Traffic',
-		subtitle: 'Traffic Widget Area',
-		style: 'composite',
-	} );
-	registry
-		.dispatch( CORE_WIDGETS )
-		.assignWidgetArea( 'TrafficArea', CONTEXT_MAIN_DASHBOARD_TRAFFIC );
-	registry.dispatch( CORE_WIDGETS ).registerWidget( 'TrafficWidget', {
-		Component: () => <div>Traffic Widget</div>,
-	} );
-	registry
-		.dispatch( CORE_WIDGETS )
-		.assignWidget( 'TrafficWidget', 'TrafficArea' );
-
-	// Content
-	registry.dispatch( CORE_WIDGETS ).registerWidgetArea( 'ContentArea', {
-		title: 'Content',
-		subtitle: 'Content Widget Area',
-		style: 'composite',
-	} );
-	registry
-		.dispatch( CORE_WIDGETS )
-		.assignWidgetArea( 'ContentArea', CONTEXT_MAIN_DASHBOARD_CONTENT );
-	registry.dispatch( CORE_WIDGETS ).registerWidget( 'ContentWidget', {
-		Component: () => <div>Content Widget</div>,
-	} );
-	registry
-		.dispatch( CORE_WIDGETS )
-		.assignWidget( 'ContentWidget', 'ContentArea' );
-
-	// Speed
-	registry.dispatch( CORE_WIDGETS ).registerWidgetArea( 'SpeedArea', {
-		title: 'Speed',
-		subtitle: 'Speed Widget Area',
-		style: 'composite',
-	} );
-	registry
-		.dispatch( CORE_WIDGETS )
-		.assignWidgetArea( 'SpeedArea', CONTEXT_MAIN_DASHBOARD_SPEED );
-	registry.dispatch( CORE_WIDGETS ).registerWidget( 'SpeedWidget', {
-		Component: () => <div>Speed Widget</div>,
-	} );
-	registry
-		.dispatch( CORE_WIDGETS )
-		.assignWidget( 'SpeedWidget', 'SpeedArea' );
-};
+import { setupDefaultChips } from './test-utils';
 
 describe( 'Dashboard Navigation', () => {
 	let previousSiteKitUserData;
